third_party/WebKit/LayoutTests/external/wpt/html/dom/dynamic-markup-insertion/opening-the-input-stream/016-expected.txt

This is a testharness.js-based test.
FAIL Timeout on original window, scopeassert_equals: Global scope from original window timeout expected 1 but got 2
PASS Timeout on original window, this object
FAIL Timeout on new window, scopeassert_equals: Global scope from original window timeout expected 1 but got 2
PASS Timeout on new window, this object
Harness: the test ran to completion.
